inline `clippy_utils::ptr` into `needless_pass_by_value`
..and make it less generic, thanks to all the callers being known
