+++ clippy_utils/src/lib.rs
@@ -63,7 +63,6 @@ pub mod mir;
 pub mod msrvs;
 pub mod numeric_literal;
 pub mod paths;
-pub mod ptr;
 pub mod qualify_min_const_fn;
 pub mod source;
 pub mod str_utils;
